+++ runtime/moonbase/tests/xcm_mock/relay_chain.rs
@@ -132,67 +132,10 @@ parameter_types! {
 	pub const MaxInstructions: u32 = 100;
 }
 
-use frame_support::ensure;
-use frame_support::traits::Contains;
-use sp_std::marker::PhantomData;
-use xcm_executor::traits::ShouldExecute;
-/// Allows execution from `origin` if it is contained in `T` (i.e. `T::Contains(origin)`) taking
-/// payments into account.
-///
-/// Only allows for `DescendOrigin` + `WithdrawAsset`, + `BuyExecution`
-pub struct AllowDescendOriginFromLocal<T>(PhantomData<T>);
-impl<T: Contains<MultiLocation>> ShouldExecute for AllowDescendOriginFromLocal<T> {
-	fn should_execute<Call>(
-		origin: &MultiLocation,
-		message: &mut Xcm<Call>,
-		max_weight: Weight,
-		_weight_credit: &mut Weight,
-	) -> Result<(), ()> {
-		log::trace!(
-			target: "xcm::barriers",
-			"AllowTopLevelPaidExecutionFromLocal origin:
-			{:?}, message: {:?}, max_weight: {:?}, weight_credit: {:?}",
-			origin, message, max_weight, _weight_credit,
-		);
-		ensure!(T::contains(origin), ());
-		let mut iter = message.0.iter_mut();
-		let i = iter.next().ok_or(())?;
-		match i {
-			DescendOrigin(..) => (),
-			_ => return Err(()),
-		}
-
-		let mut i = iter.next().ok_or(())?;
-		match i {
-			WithdrawAsset(..) => (),
-			_ => return Err(()),
-		}
-
-		i = iter.next().ok_or(())?;
-		match i {
-			BuyExecution {
-				weight_limit: Limited(ref mut weight),
-				..
-			} if *weight >= max_weight => {
-				*weight = max_weight;
-				Ok(())
-			}
-			BuyExecution {
-				ref mut weight_limit,
-				..
-			} if weight_limit == &Unlimited => {
-				*weight_limit = Limited(max_weight);
-				Ok(())
-			}
-			_ => Err(()),
-		}
-	}
-}
-
 pub type XcmRouter = super::RelayChainXcmRouter;
 pub type Barrier = (
 	TakeWeightCredit,
-	AllowDescendOriginFromLocal<Everything>,
+	xcm_primitives::AllowDescendOriginFromLocal<Everything>,
 	AllowTopLevelPaidExecutionFrom<Everything>,
 	// Expected responses are OK.
 	AllowKnownQueryResponses<XcmPallet>,
